Simplified scripts-css.
