+++ gulpfile.js
@@ -37,7 +37,6 @@ Gulp.registry(new GulpForwardReference());
     'dashboards/dist-verify',
     'dashboards/dist-zip',
     'dashboards/scripts',
-    'dashboards/scripts-css',
     'dashboards/scripts-dts',
     'dashboards/scripts-watch',
     'dashboards/test',
